Inform in the console UI when the waiting time between runs start
The last task before waiting X seconds between runs is collecting results.
If something occurs in this task without triggering an exception (e.g., the computation froze), we won't know.
This can be particularly confusing for researches and students not familiar with AR.
Sure, we can check the log file and verify the sleeping statement there, but it would be nice to have this statement also printed in the console UI.