Split Android and Trepn profilers examples
These were likely the first 2 profilers to be supported by Android Runner. At the time, it was decided to organize the examples by native and web. However, as Android Runner receive support to more profiles and additional examples were added, the examples directory organization shift to per profiler basis. Therefore, this commit updates the examples for the Trepn and Android profilers to follow this pattern.